refactor: remove deprecated v12 compatibility code and streamline achievement form handling
