+++ src/module/app/add-achievement-form.js
@@ -162,18 +162,11 @@ export class AddAchievementForm extends FormApplication {
 
   handleSelectSound(event) {
     event.preventDefault();
-    //TODO: Remove once we deprecate v12 and below
-    const isV13 = Number.parseInt(game.version.split(".")[0], 10) >= 13;
-    let fp;
-    if (isV13) {
-      // Show the foundry file picker
-      fp = new FilePicker({
-        type: "audio",
-      });
-    } else {
-      fp = new FilePicker();
-      fp.options.type = "audio";
-    }
+    // Show the foundry file picker
+    const fp = new FilePicker({
+      type: "audio",
+    });
+
     fp.render(true);
     fp.callback = (path, _filePicker) => {
       const soundInput = document.querySelector("#achievement_sound");
@@ -187,18 +180,10 @@ export class AddAchievementForm extends FormApplication {
   handleSelectImage(event) {
     event.preventDefault();
 
-    //TODO: Remove once we deprecate v12 and below
-    const isV13 = Number.parseInt(game.version.split(".")[0], 10) >= 13;
-    let fp;
-    if (isV13) {
-      // Show the foundry file picker
-      fp = new FilePicker({
-        type: "image",
-      });
-    } else {
-      fp = new FilePicker();
-      fp.options.type = "image";
-    }
+    // Show the foundry file picker
+    const fp = new FilePicker({
+      type: "image",
+    });
     fp.render(true);
     fp.callback = (path, _filePicker) => {
       const imageInput = document.querySelector("#achievement_image");
@@ -212,18 +197,11 @@ export class AddAchievementForm extends FormApplication {
   handleSelectCloakedImage(event) {
     event.preventDefault();
 
-    //TODO: Remove once we deprecate v12 and below
-    const isV13 = Number.parseInt(game.version.split(".")[0], 10) >= 13;
-    let fp;
-    if (isV13) {
-      // Show the foundry file picker
-      fp = new FilePicker({
-        type: "image",
-      });
-    } else {
-      fp = new FilePicker();
-      fp.options.type = "image";
-    }
+    // Show the foundry file picker
+    const fp = new FilePicker({
+      type: "image",
+    });
+
     fp.render(true);
     fp.callback = (path, _filePicker) => {
       const imageInput = document.querySelector("#achievement_cloaked_image");
